+++ python/guidrawer/__init__.py
@@ -1,5 +1,6 @@
 __author__ = "zebraed"
-__version__ = "3.0.0"
+__description__ = "mGear Guide Drawer Tools"
+__version__ = "3.1.0"
 __license__ = "MIT"
 __copyright__ = "Copyright 2022-2026 zebraed"